Minor formating cleanup, and tweak Sink handling
Updated some formatting issues, as well as an error in targetAngle().  Changed cvSink handling which might address issues.  Further noticed that the resolution was incorrectly set, not sure how that would have affected things but it might play some role.  Note: not compiled, just made the changes in the blind.  Will fix up later if necessary.